api/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pandas==1.3.3
python-multipart==0.0.5
pytz==2021.1
fastapi_sqlalchemy
psycopg2-binary
uuid==1.30
cryptography
"python-jose[cryptography]"
"passlib[bcrypt]"
python-decouple